feat. 일주일 치 몸무게 관리 조회 구현 중 (temp_user model height 필드 수정)

+++ lib/presentation/controller/diet_c.dart
@@ -170,7 +170,7 @@ class DietController extends GetxController {
   }
 
   // 등록 함수
-  void insert(Diet diet, int weight) async {
+  void insert(Diet diet, double weight) async {
     SharedPreferences prefs = await SharedPreferences.getInstance();
     String uuid = prefs.getString("uuid")!;
